+++ src/__tests__/cli.js
@@ -1,9 +1,7 @@
-'use strict';
-
-import ava from 'ava';
 import child from 'child_process';
 import path from 'path';
 import fs from 'fs';
+import ava from 'ava';
 
 let read = fs.readFileSync;
 let fixture = path.join(__dirname, './fixtures/nested.fixture.css');
@@ -13,14 +11,14 @@ function setup (args) {
         process.chdir(__dirname);
 
         let ps = child.spawn(process.execPath, [
-            path.resolve(__dirname, '../../bin/cmd.js')
+            path.resolve(__dirname, '../../bin/cmd.js'),
         ].concat(args));
 
         let out = '';
         let err = '';
 
-        ps.stdout.on('data', buffer => { out += buffer; });
-        ps.stderr.on('data', buffer => { err += buffer; });
+        ps.stdout.on('data', buffer => (out += buffer));
+        ps.stderr.on('data', buffer => (err += buffer));
 
         ps.on('exit', code => {
             resolve([err, out, code]);
@@ -30,23 +28,24 @@ function setup (args) {
 
 ava('cli: defaults', t => {
     return setup([fixture]).then(([err, out, code]) => {
-        t.notOk(err, 'should not error');
-        t.notOk(code, 'should exit with code 0');
-        t.same(out, read('./fixtures/nested.expanded.css', 'utf-8'), 'should transform the css');
+        t.falsy(err, 'should not error');
+        t.falsy(code, 'should exit with code 0');
+        t.deepEqual(out, read('./fixtures/nested.expanded.css', 'utf-8'), 'should transform the css');
     });
 });
 
 ava('cli: formatter', t => {
     return setup([fixture, '--format', 'compressed']).then(([err, out, code]) => {
-        t.notOk(err, 'should not error');
-        t.notOk(code, 'should exit with code 0');
-        t.same(out, read('./fixtures/nested.compressed.css', 'utf-8'), 'should transform the css');
+        t.falsy(err, 'should not error');
+        t.falsy(code, 'should exit with code 0');
+        t.deepEqual(out, read('./fixtures/nested.compressed.css', 'utf-8'), 'should transform the css');
     });
 });
 
 ava('cli: sourcemaps', t => {
     return setup([fixture, '--sourcemap']).then(([err, out]) => {
-        var hasMap = /sourceMappingURL=data:application\/json;base64/.test(out);
-        t.ok(hasMap, 'should generate a sourcemap');
+        t.falsy(err);
+        const hasMap = /sourceMappingURL=data:application\/json;base64/.test(out);
+        t.truthy(hasMap, 'should generate a sourcemap');
     });
 });