+++ src/__tests__/options.js
@@ -1,5 +1,3 @@
-"use strict";
-
 import ava from 'ava';
 import plugin from '../';
 
@@ -7,12 +5,12 @@ let tests = [{
     message: 'should have a configurable indent size',
     fixture: 'h1{color:black}',
     expected: 'h1 {\n  color: black;\n}\n',
-    options: {indentSize: 2}
+    options: {indentSize: 2},
 }, {
     message: 'should allow disabling of the cascade',
     fixture: 'h1{-webkit-border-radius: 12px; border-radius: 12px; }',
     expected: 'h1 {\n    -webkit-border-radius: 12px;\n    border-radius: 12px;\n}\n',
-    options: {cascade: false}
+    options: {cascade: false},
 }];
 
 function perfectionist (css, options) {
@@ -22,6 +20,6 @@ function perfectionist (css, options) {
 ava('perfectionist options', (t) => {
     tests.forEach(test => {
         let options = test.options || {};
-        t.same(perfectionist(test.fixture, options), test.expected);
+        t.deepEqual(perfectionist(test.fixture, options), test.expected);
     });
 });