+++ src/maxSelectorLength.js
@@ -1,30 +1,29 @@
-'use strict';
-
 import {list} from 'postcss';
 import space from './space';
 import getIndent from './getIndent';
-import assign from 'object-assign';
 
 function splitProperty (rule, prop, opts) {
-    opts = assign({
-        reindent: false
-    }, opts);
-    let max = opts.max;
-    let property = rule[prop];
-    if (!max || !property) { return; }
-    let exploded = list.comma(property);
-    if (property.length > max || opts.reduce) {
+    const {breakEvery, reindent, reduce, max} = {
+        reindent: false,
+        ...opts,
+    };
+    const property = rule[prop];
+    if (!max || !property) {
+        return;
+    }
+    const exploded = list.comma(property);
+    if (property.length > max || reduce) {
         let indent = 0;
-        if (typeof opts.reindent === 'function') {
-            indent = opts.reindent(rule);
+        if (typeof reindent === 'function') {
+            indent = reindent(rule);
         }
         rule[prop] = exploded.reduce((lines, chunk) => {
-            if (opts.breakEvery) {
+            if (breakEvery) {
                 lines.push(chunk);
                 return lines;
             }
             if (lines[lines.length - 1].length + indent <= max) {
-                let merged = lines[lines.length - 1] + ', ' + chunk;
+                const merged = `${lines[lines.length - 1]}, ${chunk}`;
                 if (indent + merged.length <= max) {
                     lines[lines.length - 1] = merged;
                     return lines;
@@ -38,11 +37,11 @@ function splitProperty (rule, prop, opts) {
 
 export function maxAtRuleLength (rule, {maxAtRuleLength: max}) {
     return splitProperty(rule, 'params', {
-        max: max,
+        max,
         breakEvery: true,
-        reindent: function (rule) {
-            return rule.name.length + 2;
-        }
+        reindent: function (r) {
+            return r.name.length + 2;
+        },
     });
 }
 
@@ -50,9 +49,9 @@ export function maxSelectorLength (rule, opts) {
     return splitProperty(rule, 'selector', {
         max: opts.maxSelectorLength,
         reduce: true, // where possible reduce to one line
-        reindent: function (rule) {
-            return getIndent(rule, opts.indentSize).length;
-        }
+        reindent: function (r) {
+            return getIndent(r, opts.indentSize).length;
+        },
     });
 }
 
@@ -61,10 +60,10 @@ export function maxValueLength (rule, {maxValueLength: max}) {
         rule.value = rule.raws.value.raw;
     }
     return splitProperty(rule, 'value', {
-        max: max,
+        max,
         breakEvery: true,
-        reindent: function (rule) {
-            return getIndent(rule).length + rule.prop.length + 2;
-        }
+        reindent: function (r) {
+            return getIndent(r).length + r.prop.length + 2;
+        },
     });
 }