+++ src/sameLine.js
@@ -1,3 +1,1 @@
-'use strict';
-
 export default (a, b) => a.source.end.line === b.source.start.line;